+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changes
 =======
 
+## UNRELEASED
+
+* Add debug log message if no service or function `package` configurations apply.
+
 ## 0.1.0
 
 * Initial release.